Use methods for tus views
